+++ packages/router/tests/route.spec.ts
@@ -842,7 +842,7 @@ describe("routes", () => {
           });
         });
 
-        describe("status", () => {
+        describe("meta", () => {
           it("is undefined if not set by route.response()", () => {
             const routes = prepareRoutes({
               routes: [
@@ -862,10 +862,10 @@ describe("routes", () => {
               }
             });
             const { response } = router.current();
-            expect(response.status).toBeUndefined();
+            expect(response.meta).toBeUndefined();
           });
 
-          it("is the status value of object returned by route.response()", () => {
+          it("is the meta value of object returned by route.response()", () => {
             const routes = prepareRoutes({
               routes: [
                 {
@@ -873,7 +873,10 @@ describe("routes", () => {
                   path: "",
                   response: () => {
                     return {
-                      status: 451
+                      meta: {
+                        title: "A Route",
+                        status: 451
+                      }
                     };
                   }
                 }
@@ -885,7 +888,10 @@ describe("routes", () => {
               }
             });
             const { response } = router.current();
-            expect(response.status).toBe(451);
+            expect(response.meta).toMatchObject({
+              title: "A Route",
+              status: 451
+            });
           });
         });
 
@@ -931,51 +937,6 @@ describe("routes", () => {
             });
             const { response } = router.current();
             expect(response.data).toMatchObject({ test: "value" });
-          });
-        });
-
-        describe("title", () => {
-          it("is undefined if not set by route.response()", () => {
-            const routes = prepareRoutes({
-              routes: [
-                {
-                  name: "State",
-                  path: ":state"
-                }
-              ]
-            });
-            const router = createRouter(inMemory, routes, {
-              history: {
-                locations: ["/AZ"]
-              }
-            });
-
-            const { response } = router.current();
-            expect(response.title).toBeUndefined();
-          });
-
-          it("is the title value of the object returned by route.response()", () => {
-            const routes = prepareRoutes({
-              routes: [
-                {
-                  name: "State",
-                  path: ":state",
-                  response: () => {
-                    return {
-                      title: "A State"
-                    };
-                  }
-                }
-              ]
-            });
-            const router = createRouter(inMemory, routes, {
-              history: {
-                locations: ["/VA"]
-              }
-            });
-
-            const { response } = router.current();
-            expect(response.title).toBe("A State");
           });
         });
 
@@ -1171,44 +1132,6 @@ describe("routes", () => {
           });
         });
 
-        describe("error", () => {
-          it("is undefined for good responses", () => {
-            const routes = prepareRoutes({
-              routes: [{ name: "Contact", path: "contact" }]
-            });
-            const router = createRouter(inMemory, routes, {
-              history: {
-                locations: ["/contact"]
-              }
-            });
-            const { response } = router.current();
-            expect(response.error).toBeUndefined();
-          });
-
-          it("is the error value on the object returned by route.response()", () => {
-            const routes = prepareRoutes({
-              routes: [
-                {
-                  name: "A Route",
-                  path: "",
-                  response: () => {
-                    return {
-                      error: "woops"
-                    };
-                  }
-                }
-              ]
-            });
-            const router = createRouter(inMemory, routes, {
-              history: {
-                locations: ["/"]
-              }
-            });
-            const { response } = router.current();
-            expect(response.error).toBe("woops");
-          });
-        });
-
         describe("redirect", () => {
           it("contains the expected properties", () => {
             const routes = prepareRoutes({
@@ -1329,7 +1252,7 @@ describe("routes", () => {
                   response: () => {
                     return {
                       body: "body",
-                      status: undefined
+                      meta: undefined
                     };
                   }
                 }
@@ -1339,7 +1262,7 @@ describe("routes", () => {
             const { response } = router.current();
             expect(response.name).toBe("A Route");
             expect(response.hasOwnProperty("body")).toBe(true);
-            expect(response.hasOwnProperty("status")).toBe(false);
+            expect(response.hasOwnProperty("meta")).toBe(false);
           });
         });